+++ demo/FormDemo.js
@@ -40,26 +40,26 @@ class Demo extends React.Component {
         super(props);
         this.state = {
             jsxvalues: {
-                test1: "我是测试",
-                fruit: "apple",
-                city: "nj",
-                textArea: "我是多行文本",
-                date: "2015-09-01",
-                checkbox: ["sea"],
-                dicts:{
-                   datas:[
-                      {
-                        city: "hz",
-                        email: "333",
-                        name: "33"
-                      }, {
-                        city: "bj",
-                        email: "33322",
-                        name: "3322"
-                      }
-                   ]
-                },
-                cascade: ["a", "ab"]
+                // test1: "我是测试",
+                // fruit: "apple",
+                // city: "nj",
+                // textArea: "我是多行文本",
+                // date: "2015-09-01",
+                // checkbox: ["sea"],
+                // dicts:{
+                //    datas:[
+                //       {
+                //         city: "hz",
+                //         email: "333",
+                //         name: "33"
+                //       }, {
+                //         city: "bj",
+                //         email: "33322",
+                //         name: "3322"
+                //       }
+                //    ]
+                // },
+                // cascade: ["a", "ab"]
             }
         }
     }
@@ -205,7 +205,7 @@ class Demo extends React.Component {
 
         return (
             <div>
-                <Form ref="form" jsxmode={Constants.MODE.VIEW} jsxvalues={me.state.jsxvalues} jsxonChange={me.handleChange.bind(me)}>
+                <Form ref="form" jsxmode={Constants.MODE.EDIT} jsxvalues={me.state.jsxvalues} jsxonChange={me.handleChange.bind(me)}>
                     <FormRowTitle jsxtitle="我是行标题"/>
                     <FormRow>
                         <InputFormField
@@ -298,8 +298,8 @@ class Demo extends React.Component {
                         width: 200
 
                      }}/>
-                    {/*<EditorFormField jsxname="editor"
-                                     jsxlabel="富文本编辑器"/>*/}
+                    <EditorFormField jsxname="editor"
+                                     jsxlabel="富文本编辑器"/>
 
                     {/*<GridFormField jsxname="dicts" jsxlabel="薪酬字典" {...renderProps}>
                     </GridFormField>*/}