+++ demo/FormDemo.js
@@ -61,6 +61,12 @@ class Demo extends React.Component {
                 },
                 cascade: ["a", "ab"],
                 editor: "a"
+            },
+            jsxdata: {
+                "bj": "北京",
+                "nj": "南京",
+                "dj": "东京",
+                "xj": "西京"
             }
         }
     }
@@ -91,6 +97,12 @@ class Demo extends React.Component {
                 date: "2015-09-01",
                 checkbox: ["sea"],
                 // cascade: ["a", "ab"]
+            },
+            jsxdata: {
+                "bj": "北",
+                "nj": "南",
+                "dj": "东",
+                "xj": "西"
             }
         })
     }
@@ -205,7 +217,7 @@ class Demo extends React.Component {
 
 
         return (
-            <div>
+            <div className="demo">
                 <Form ref="form" jsxmode={Constants.MODE.EDIT} jsxvalues={me.state.jsxvalues} jsxonChange={me.handleChange.bind(me)}>
                     <FormRowTitle jsxtitle="我是行标题"/>
                     <FormRow>
@@ -215,38 +227,6 @@ class Demo extends React.Component {
                          jsxmode={Constants.MODE.VIEW}
                          jsxtips="请输入数字"
                          jsxrules={{validator: Validators.isNotEmpty, errMsg: "不能为空"}}/>
-                        <RadioGroupFormField jsxname="fruit" jsxlabel="Radio" jsxflex={1}>
-                                <RadioItem value="apple" text="Apple"/>
-                                <RadioItem value="orange" text="Orange"/>
-                                <RadioItem value="watermelon" text="Watermelon"/>
-                        </RadioGroupFormField>
-                        <ButtonGroupFormField jsxshow={false}>
-                            <Button size="medium" type="submit" onClick={me.handleFormClick.bind(me)}>提交</Button>
-                            <Button size="medium" type="reset">取消</Button>
-                        </ButtonGroupFormField>
-                    </FormRow>
-                    <FormRow>
-                        <SelectFormField
-                         jsxlabel="单选"
-                         jsxname="city"
-                         jsxfetchUrl="http://suggest.taobao.com/sug"
-                         afterFetch={(obj) => {
-                            let data = {};
-                            obj.result.forEach((item, index) => {
-                                data[item[1]] = item[0];
-                            });
-                            console.log(data);
-                            return data;
-                         }}
-                         jsxdata={{
-                            "bj": "北京",
-                            "nj": "南京",
-                            "dj": "东京",
-                            "xj": "西京"
-                         }}
-                         jsxstyle={{
-                            width: 200
-                         }}/>
                          <NumberInputFormField
                           jsxname="number"
                           jsxlabel="数字输入框"
@@ -257,6 +237,39 @@ class Demo extends React.Component {
                             {validator: Validators.isNotEmpty, errMsg: "不能为空"},
                             {validator: Validators.isNum, errMsg: "请输入数字"}
                          ]}/>
+                        
+                        <ButtonGroupFormField jsxshow={false}>
+                            <Button size="medium" type="submit" onClick={me.handleFormClick.bind(me)}>提交</Button>
+                            <Button size="medium" type="reset">取消</Button>
+                        </ButtonGroupFormField>
+                    </FormRow>
+                    <FormRow>
+                        <RadioGroupFormField jsxname="fruit" jsxlabel="Radio" jsxflex={1}>
+                                <RadioItem value="apple" text="Apple"/>
+                                <RadioItem value="orange" text="Orange"/>
+                                <RadioItem value="watermelon" text="Watermelon"/>
+                        </RadioGroupFormField>
+                        <CheckboxGroupFormField jsxname="checkbox" jsxlabel="复选框">
+                          <CheckboxItem value="air" text="天空"/>
+                          <CheckboxItem value="sea" text="大海"/>
+                        </CheckboxGroupFormField>
+                    </FormRow>
+                    <TextAreaFormField jsxname="textArea" jsxlabel="多行文本框"/>
+                    <FormRowTitle jsxtitle="我是行标题2"/>
+                    <FormRow>
+                        <SelectFormField
+                         jsxlabel="单选"
+                         jsxname="city"
+                         jsxfetchUrl="http://suggest.taobao.com/sug"
+                         afterFetch={(obj) => {
+                            let data = {};
+                            obj.result.forEach((item, index) => {
+                                data[item[1]] = item[0];
+                            });
+                            return data;
+                         }}
+                         jsxdata={me.state.jsxdata}/>
+                        <DateFormField jsxname="date" jsxlabel="日期" jsxfrom="2015-10-2" jsxto="2015-10-10"/>
                     </FormRow>
                     <FormRow>
                         <SelectFormField
@@ -270,20 +283,21 @@ class Demo extends React.Component {
                                 data[item[1]] = item[0];
                             });
                             return data;
-                         }}
-                         jsxstyle={{width: 200}}/>
-
+                         }}/>
+                         <SelectFormField
+                         jsxlabel="多选模式"
+                         jsxname="goods2"
+                         jsxmultiple={true}
+                         jsxfetchUrl="http://suggest.taobao.com/sug"
+                         afterFetch={(obj) => {
+                            let data = {};
+                            obj.result.forEach((item, index) => {
+                                data[item[1]] = item[0];
+                            });
+                            return data;
+                         }}/>
                     </FormRow>
-                    <FormRowTitle jsxtitle="我是行标题2"/>
                     <FormRow>
-                        <TextAreaFormField jsxname="textArea" jsxlabel="多行文本框"/>
-                        <CheckboxGroupFormField jsxname="checkbox" jsxlabel="复选框">
-                          <CheckboxItem value="air" text="天空"/>
-                          <CheckboxItem value="sea" text="大海"/>
-                        </CheckboxGroupFormField>
-                    </FormRow>
-                    <FormRow>
-                        <DateFormField jsxname="date" jsxlabel="日期" jsxfrom="2015-10-2" jsxto="2015-10-10"/>
                         <UploadFormField
                           jsxname="upload"
                           jsxlabel="上传"
@@ -294,11 +308,7 @@ class Demo extends React.Component {
                     <CascadeSelectFormField
                      jsxdata={casData}
                      jsxname="cascade"
-                     jsxlabel="级联选择"
-                     jsxstyle={{
-                        width: 200
-
-                     }}/>
+                     jsxlabel="级联选择"/>
                     <EditorFormField jsxname="editor"
                                      jsxlabel="富文本编辑器"
                                      jsxcontent="1"/>
@@ -306,9 +316,9 @@ class Demo extends React.Component {
                     <GridFormField jsxname="dicts" jsxlabel="薪酬字典" {...renderProps}>
                     </GridFormField>
                     <ButtonGroupFormField>
-                        <Button size="medium" type="submit" onClick={me.handleFormClick.bind(me)}>提交</Button>
-                        <Button size="medium" type="reset">取消</Button>
-                        <Button size="medium" onClick={me.handleValueChange.bind(me)}>修改 props</Button>
+                        <Button size="medium" action="submit" onClick={me.handleFormClick.bind(me)}>提交</Button>
+                        <Button size="medium" type="secondary" action="reset">取消</Button>
+                        <Button size="medium" type="secondary" onClick={me.handleValueChange.bind(me)}>修改 props</Button>
                     </ButtonGroupFormField>
                 </Form>
             </div>