+++ HISTORY.md
@@ -1,3 +1,10 @@
+## 1.0.7
+
+`NEW` FormField can set its own mode different from Form now.
+`NEW` add GridFormField.
+`FIXED` when props jsxvalues change, form cannot handle correctly.
+
+
 ## 1.0.6
 
 `NEW` add props jsxshowLabel